requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.20
asyncpg==0.31.0
redis==5.0.1
pydantic==2.7.4
pytest==7.4.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-dotenv==1.0.1